[Refactor] ShippingInfo => Shipping
- 준규님께서 ShippingInfo 를 별도로 정의해서 사용하던 것을 공통 필드만 Shipping으로 묶어서 관리하도록 재설계
