+++ src/main/java/com/bbangle/bbangle/delivery/domain/Shipping.java
@@ -2,9 +2,7 @@ package com.bbangle.bbangle.delivery.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
-import jakarta.persistence.Embedded;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
+import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,13 +12,18 @@ import lombok.NoArgsConstructor;
 @Embeddable
 public class Shipping {
 
-    @Embedded
-    private ShippingInfo shippingInfo;
+    @Column(length = 255)
+    private String deliveryMemo;
 
-    @Column(name = "status", length = 30, columnDefinition = "varchar(30)")
-    @Enumerated(EnumType.STRING)
-    private ShippingStatus status;
+    @Column(length = 50)
+    private String courierName;
 
+    @Column(length = 50)
+    private String trackingNumber;
 
+    private Integer fee;
+
+    private LocalDateTime shippedAt;
+    private LocalDateTime deliveredAt;
 
 }
