+++ src/main/java/com/bbangle/bbangle/order/domain/OrderDelivery.java
@@ -3,12 +3,11 @@ package com.bbangle.bbangle.order.domain;
 import com.bbangle.bbangle.common.domain.BaseEntity;
 import com.bbangle.bbangle.delivery.domain.Receiver;
 import com.bbangle.bbangle.delivery.domain.Sender;
-import com.bbangle.bbangle.delivery.domain.ShippingInfo;
+import com.bbangle.bbangle.delivery.domain.Shipping;
 import com.bbangle.bbangle.order.domain.model.OrderDeliveryStatus;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
@@ -21,7 +20,6 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
@@ -40,7 +38,7 @@ public class OrderDelivery extends BaseEntity {
     private Receiver receiver;
 
     @Embedded
-    private ShippingInfo shippingInfo;
+    private Shipping shipping;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "delivery_status", columnDefinition = "VARCHAR(20)")
@@ -49,4 +47,5 @@ public class OrderDelivery extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "order_item_id")
     private OrderItem orderItem;
+
 }
